requirements.txt
amadeus==5.2.0
aniso8601==9.0.1
CacheControl==0.12.10
cachetools==5.0.0
certifi==2020.6.20
charset-normalizer==2.0.11
click==8.0.3
colorama==0.4.4
firebase-admin==5.2.0
Flask==2.0.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
google-api-core==2.5.0
google-api-python-client==2.36.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.1.0
googleapis-common-protos==1.54.0
grpcio==1.43.0
grpcio-status==1.43.0
httplib2==0.20.4
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.3
packaging==21.3
proto-plus==1.19.9
protobuf==3.19.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
rsa==4.8
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.2
wincertstore==0.2
